handle int constants wider than 32 bits
Integers in the form of a based literal constant
are processed correctly.
Integers in the form of a simple decimal number
that exceed 32 bits are mishandled,
and limited to 32 bits as a result.
Now they will be stored as a string and
processed the same way as based literal constants.

Signed-off-by: Wojciech Sipak <wsipak@antmicro.com>